# docs/api/openapi.yaml
components:
  schemas:
...
    api.SwaggerCreatePageRequest:
      properties:
        content:
          example: |-
            # Welcome
            This is a wiki page.
          type: string
        parent_id:
          type: string
        position:
          example: 0
          type: integer
        title:
          example: Getting Started
          type: string
      type: object
...
    api.SwaggerErrorDetail:
      properties:
        code:
          example: UNAUTHORIZED
          type: string
        message:
          example: invalid email or password
          type: string
        request_id:
          example: req_a8d9912a
          type: string
      type: object
    api.SwaggerErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/api.SwaggerErrorDetail'
      type: object
...
    api.SwaggerMessageResponse:
      properties:
        message:
          example: operation completed
          type: string
      type: object
    api.SwaggerMovePageRequest:
      properties:
        parent_id:
          type: string
        position:
          example: 1
          type: integer
      type: object
...
    api.SwaggerUpdatePageRequest:
      properties:
        content:
          example: '# Updated Content'
          type: string
        expected_version:
          example: 1
          type: integer
        title:
          example: Getting Started (updated)
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      description: 'JWT Bearer token. Format: "Bearer <token>". Obtain via POST /api/v1/auth/login'
      in: header
      name: Authorization
      type: apiKey
...
info:
  contact:
    email: hello@azimuthalhq.com
    name: Azimuthal HQ
    url: https://azimuthalhq.com
  description: |-
    The fully open-source alternative to Jira, Confluence, and Jira Service Desk.
    Self-hostable, single binary, Apache 2.0 licensed.

    ## Authentication
    All endpoints except /auth/login and /auth/register require a Bearer JWT token.
    Obtain a token via POST /api/v1/auth/login, then include it as:
    `Authorization: Bearer <your-token>`
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Azimuthal API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /spaces/{spaceID}/wiki:
    get:
      description: Returns all pages in the specified space.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  additionalProperties: {}
                  type: object
                type: array
          description: List of pages
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid space ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: List wiki pages
      tags:
      - wiki
    post:
      description: Creates a new wiki page. Author is set from the JWT.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SwaggerCreatePageRequest'
                description: Page details
                summary: body
        description: Page details
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Created page
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Create wiki page
      tags:
      - wiki
  /spaces/{spaceID}/wiki/{pageID}:
    delete:
      description: Soft-deletes a wiki page by ID.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: Page ID (UUID)
        in: path
        name: pageID
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Deleted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Delete wiki page
      tags:
      - wiki
    get:
      description: Returns a single wiki page by ID.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: Page ID (UUID)
        in: path
        name: pageID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Page details
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid ID
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Get wiki page
      tags:
      - wiki
    put:
      description: Updates a page with optimistic locking. Returns 409 with conflict
        details if version mismatch.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: Page ID (UUID)
        in: path
        name: pageID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SwaggerUpdatePageRequest'
                description: Updated fields
                summary: body
        description: Updated fields
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Updated page
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Validation error
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
        "409":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Version conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Update wiki page
      tags:
      - wiki
  /spaces/{spaceID}/wiki/{pageID}/diff:
    get:
      description: Returns the diff between two page versions. Requires 'from' and
        'to' query parameters.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: Page ID (UUID)
        in: path
        name: pageID
        required: true
        schema:
          type: string
      - description: From version number
        in: query
        name: from
        required: true
        schema:
          type: integer
      - description: To version number
        in: query
        name: to
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Diff result
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Missing or invalid params
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Diff page versions
      tags:
      - wiki
  /spaces/{spaceID}/wiki/{pageID}/move:
    post:
      description: Changes a page's parent or position in the tree hierarchy.
      parameters:
      - description: Space ID (UUID)
        in: path
        name: spaceID
        required: true
        schema:
          type: string
      - description: Page ID (UUID)
        in: path
        name: pageID
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.SwaggerMovePageRequest'
                description: New position
                summary: body
        description: New position
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerMessageResponse'
          description: Page moved
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Invalid request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not authenticated
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.SwaggerErrorResponse'
          description: Internal error
      security:
      - BearerAuth: []
      summary: Move wiki page
      tags:
      - wiki
...
servers:
- url: localhost:8080/api/v1